+++ app/(routes)/(site)/blog/[slug]/page.tsx
@@ -1,7 +1,7 @@
-import { fetchPost } from '@/app/lib/actions';
 import Post from '@/app/ui/posts/Post';
 import ScrollBack from '@/app/ui/common/ScrollBack';
 import { Metadata } from 'next';
+import { selectPosts } from '@/app/db/queries';
 
 type PostProps = {
   params: Promise<{ slug: string }>;
@@ -11,16 +11,12 @@ export async function generateMetadata({
   params,
 }: PostProps): Promise<Metadata> {
   const slug = (await params).slug;
-  const result = await fetchPost(slug);
+  const post = (await selectPosts({ slug }))[0];
 
-  if (result.success) {
-    return {
-      title: result.data.data.title,
-      description: result.data.excerpt,
-    };
-  }
-
-  return {};
+  return {
+    title: post.title,
+    description: post.excerpt,
+  };
 }
 
 export default async function PostPage({