+++ app/(routes)/writr/posts/page.tsx
@@ -1,14 +1,9 @@
-import { fetchAllPosts } from '@/app/lib/actions';
-import { notFound } from 'next/navigation';
 import AdminPostPreview from '@/app/ui/posts/PostAdminPreview';
 import Header from '@/app/ui/common/Header';
-import PlaceholderPage from '@/app/ui/common/PlaceholderPage';
+import { selectPosts } from '@/app/db/queries';
 
 export default async function Page() {
-  const posts = await fetchAllPosts({ publishedOnly: false });
-
-  if (!posts.success) return notFound();
-  if (posts.data.length === 0) return <PlaceholderPage />;
+  const rows = await selectPosts({});
 
   return (
     <>
@@ -16,7 +11,7 @@ export default async function Page() {
         <h1>All Posts</h1>
       </Header>
       <section>
-        {posts.data.map((post, index) => (
+        {rows.map((post, index) => (
           <AdminPostPreview key={index} post={post} />
         ))}
       </section>